+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Support new mapping syntax for Leo
+
 ## [0.1.2]
 
 - Fixed function call highlighting
@@ -21,12 +23,14 @@
 - Fixed some highlighting and formatting issues for Leo
 
 ## [0.0.2]
+
 - Added basic code completion
 - Added basic formatting rules
 - Implemented structure view
 - Implemented go to symbol
 
 ## [0.0.1-3]
+
 - Added code folding support
 - Added parse recovering
 - Reference circuit components in circuit expressions